project/requirements.txt
transformers>=4.35.0,<5.0.0
torch>=2.0.0,<2.2.0
datasets>=2.14.0,<3.0.0
accelerate>=0.24.0,<1.0.0
Pillow>=10.0.0,<11.0.0
numpy>=1.24.0,<2.0.0
seqeval>=1.2.2
scikit-learn>=1.3.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0